move calc_fft_localy to tools
